Nullability and comprehensive code checks - Validated
- Full coverage for @Nullable and @NonNull annotations for all methods
  - Explicit code checks for null are retained for backwards compatibility (code inspection warnings are suppressed)
- Full coverage for "final" method parameters and constants
- Full review and update/suppression of all code inspection warnings
- Validated on Pixel3, Samsung A20 with iPhone X (14.2) and Denmark

+++ herald/src/test/java/io/heraldprox/herald/sensor/datatype/Int64Tests.java
@@ -24,8 +24,8 @@ public class Int64Tests {
         // Zero
         assertEquals(0, new Int64(0).value);
         // Clamping
-        assertEquals(Int64.min.value, new Int64((long) Int64.min.value).value);
-        assertEquals(Int64.max.value, new Int64((long) Int64.max.value).value);
+        assertEquals(Int64.min.value, new Int64(Int64.min.value).value);
+        assertEquals(Int64.max.value, new Int64(Int64.max.value).value);
         // String
         assertNotNull(Int64.min.toString());
         assertNotNull(Int64.max.toString());
